Started UI: Login Page and Database Class
I created the main class which will only instantiate the classes and run the program.

I added the first UI page which was the login page which is under the "Login" class. Handles entry of credentials to confirm if the user should be allowed to access the tracker.

Started work on the "Database" class of the project. The "Login" class credentials confirms with the "Database" if the entered ID and password exists within the MySQL database.